contracts-llm/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
pdfplumber==0.11.4
pdfminer.six==20231228
pillow==10.4.0
rapidfuzz==3.9.6
sentence-transformers==3.0.1
httpx==0.28.1
python-dotenv==1.0.1
